+++ srcs/builtins/ft_unset.c
@@ -2,7 +2,7 @@
 
 //t_env	*g_sh_var.environ;
 
-bool	is_valid_unset_args(char *str)
+bool	is_valid_unset_args(char *str) //same is_valid_vat
 {
 	int	i;
