+++ srcs/execution/execute.c
@@ -9,13 +9,6 @@ void	error_checker(char *msg, int n)
 	}
 }
 
-void	err_is_a_directory(char *filename)
-{
-	ft_putstr_fd("minishell: ", 2);
-	ft_putstr_fd(filename, 2);
-	ft_putstr_fd(": is a directory\n", 2);
-}
-
 void	xunlink(const char *pathname)
 {
 	int	ret;
@@ -100,6 +93,7 @@ bool	which_builtin(char **cmd)
 	return (false);
 }
 
+//
 size_t	count_envlist(t_env *env)
 {
 	t_env	*head;
@@ -139,26 +133,9 @@ char	**envlist_to_str(t_env *env)
 	envstr[++i] = NULL;
 	return (envstr);
 }
+//
 
-int	exec_builtin(t_cmd *cmd)
-{
-	const char	*builtins[] = {"cd", "echo", "unset", \
-		"export", "exit", "pwd", "env"};
-	static int	(*builtin_func[])(char **) = {&ft_cd, &ft_echo, &ft_unset, \
-		&ft_export, &ft_exit, &ft_pwd, &ft_env};
-	size_t		i;
-
-	i = 0;
-	while (i < (sizeof(builtins) / sizeof(builtins[0])))
-	{
-		if (ft_strncmp(cmd->cmd[0], builtins[i], ft_strlen(builtins[i]) + 1)
-			== 0)
-			return ((*builtin_func[i])(cmd->cmd));
-		i++;
-	}
-	return (1);
-}
-
+//
 char	*create_and_check_path(char *env_path, char *cmd_name)
 {
 	char	*join_path;
@@ -228,34 +205,13 @@ char	*check_path(char *path)
 		errno = EACCES;
 	return (join_path);
 }
+//
 
 bool	is_path(char *cmd_name)
 {
 	if (ft_strchr(cmd_name, '/'))
 		return (true);
 	return (false);
-}
-
-void	print_exec_process_error(char *cmd, char *msg, int status)
-{
-	ft_putstr_fd("minishell: ", 2);
-	ft_putstr_fd(cmd, 2);
-	ft_putstr_fd(msg, 2);
-	ft_putstr_fd("\n", 2);
-	if (status == EACCES)
-		exit(126);
-	else if (status == ENOENT || status == -1)
-		exit(127);
-}
-
-void	print_access_err(char *msg)
-{
-	if (errno == EACCES)
-		print_exec_process_error(msg, ": Permission denied", EACCES);
-	else if (errno == ENOENT && !is_path(msg))
-		print_exec_process_error(msg, ": command not found", ENOENT);
-	else
-		print_exec_process_error(msg, ": No such file or directory", -1);
 }
 
 bool	is_directory(char *pathname)
@@ -306,179 +262,6 @@ void	exec_by_cmd_name(t_cmd *cmd, char **envstr)
 //		exit(execve(path, cmd->cmd, envstr));
 	execve(path, cmd->cmd, envstr);
 	err_exit("execve error: ");
-}
-
-void	exec_others(t_cmd *cmd)
-{
-	char	**envstr;
-
-	envstr = envlist_to_str(g_sh_var.environ);
-	if (is_path(cmd->cmd[0]))
-	{
-		exec_by_cmd_path(cmd, envstr);
-	}
-	else
-	{
-		exec_by_cmd_name(cmd, envstr);
-	}
-}
-
-int	print_redirect_err(char *file_name)
-{
-	ft_putstr_fd("minishell: ", 2);
-	ft_putstr_fd(file_name, 2);
-	if (errno == EACCES)
-		ft_putstr_fd(": Permission denied\n", 2);
-	else if (errno == ENOENT)
-		ft_putstr_fd(": No such file or directory\n", 2);
-	else
-		ft_putstr_fd(": ambiguous redirect\n", 2);
-	return (0);
-//	return (1);
-}
-
-bool	is_valid_file_name(t_redirect *redirect)
-{
-	if (redirect->type == REDIRECT_NONE)
-		return (print_redirect_err(redirect->file_name));
-	else if (redirect->type == APPEND || redirect->type == REDIRECT_OUT)
-	{
-		if ((access(redirect->file_name, W_OK) && errno == EACCES)
-			|| redirect->file_name[0] == '\0')
-			return (print_redirect_err(redirect->file_name));
-		else if (is_directory(redirect->file_name))
-		{
-			//ft_putstr_fd("minishell: ", 2);
-			//ft_putstr_fd(redirect->file_name, 2);
-			//ft_putstr_fd(": is a directory\n", 2);
-			err_is_a_directory(redirect->file_name);
-			return (0);
-		}
-	}
-	else
-	{
-		if (access(redirect->file_name, R_OK))
-			return (print_redirect_err(redirect->file_name));
-	}
-	return (1);
-}
-
-int	open_and_dup2(t_redirect *redirect)
-{
-	int	fd;
-
-	if (!is_valid_file_name(redirect))
-		return (1);
-	if (redirect->type == APPEND)
-	{
-		fd = open(redirect->file_name, O_CREAT | O_WRONLY | O_APPEND, 0644);
-		xdup2(fd, 1);
-	}
-	else if (redirect->type == REDIRECT_OUT)
-	{
-		fd = open(redirect->file_name, O_CREAT | O_WRONLY | O_TRUNC, 0644);
-		xdup2(fd, 1);
-	}
-	else
-	{
-		fd = open(redirect->file_name, O_RDONLY);
-		xdup2(fd, 0);
-		if (redirect->type == HEREDOC)
-			xunlink(redirect->file_name);
-	}
-	xclose(fd);
-	return (0);
-}
-
-void	create_heredoc_tmpfile(t_redirect *redirect_in)
-{
-	int		fd;
-
-//	flag = O_CREAT | O_WRONLY | O_TRUNC;
-	fd = open(redirect_in->file_name, O_CREAT | O_WRONLY | O_TRUNC, 00644);
-	ft_putstr_fd(redirect_in->documents, fd);
-	free(redirect_in->documents);
-	redirect_in->documents = NULL;
-	xclose(fd);
-}
-
-int	do_redirect(t_cmd *cmd)
-{
-	t_redirect	*redirect_in;
-	t_redirect	*redirect_out;
-
-	redirect_in = cmd->redirect_in;
-	while (redirect_in->next)
-	{
-		redirect_in = redirect_in->next;
-		if (redirect_in->type == HEREDOC)
-		{
-			create_heredoc_tmpfile(redirect_in);
-		}
-		if (open_and_dup2(redirect_in))
-			return (1);
-	}
-	redirect_out = cmd->redirect_out;
-	while (redirect_out->next)
-	{
-		redirect_out = redirect_out->next;
-		if (open_and_dup2(redirect_out))
-			return (1);
-	}
-	return (0);
-}
-
-int	exe_process(t_cmd *cmd)
-{
-	int	status;
-
-	errno = 0;
-	status = 0;
-	if (is_redirect(cmd))
-	{
-		if (do_redirect(cmd))
-		{
-			return (1);
-		}
-	}
-	if (which_builtin(cmd->cmd))
-		status = exec_builtin(cmd);
-	else if (cmd->cmd[0] != NULL && cmd->cmd[0][0] != '\0')
-		exec_others(cmd);
-	return (status);
-}
-
-static void	sig_handler(int sig)
-{
-	(void)sig;
-	printf("\n");
-	rl_on_new_line();
-	rl_replace_line("", 0);
-}
-
-pid_t	exe_cmd(t_cmd *cmd, int pipe_flag)
-{
-	int		fd[2];
-	pid_t	pid;
-
-	xpipe(fd);
-	set_signal_handler(SIGINT, sig_handler);
-	pid = xfork();
-	if (pid == 0)
-	{
-		set_signal_handler(SIGINT, SIG_DFL);
-		set_signal_handler(SIGQUIT, SIG_DFL);
-		if (pipe_flag != 1)
-			exit(exe_process(cmd));
-		xdup2(fd[1], 1);
-		xclose(fd[1]);
-		xclose(fd[0]);
-		exit(exe_process(cmd));
-	}
-	xdup2(fd[0], 0);
-	xclose(fd[1]);
-	xclose(fd[0]);
-	return (pid);
 }
 
 pid_t	exe_terminal_node(t_node *node, int pipe_flag) //t_cmd *cmd is better?
