+++ srcs/lexer/lexer_utils.c
@@ -1,7 +1,7 @@
 #include "minishell.h"
 
-void	ft_exit2(char *msg)
-{
-	printf("%s\n", msg);
-	exit(EXIT_FAILURE);
-}
+//void	ft_exit2(char *msg)
+//{
+//	printf("%s\n", msg);
+//	exit(EXIT_FAILURE);
+//}
